+++ CHANGES.md
@@ -15,6 +15,7 @@ v1.2.0 (Month DD, YYYY)
 - Updated the token reading code to protect against some obvious abuses of the
   PDF format.
 - Updated the xref reading code to protect against loops.
+- Fixed handling of encrypted PDFs with per-object file IDs (Issue #42)
 - Fixed handling of of trailer dictionaries that started immediately after the
   "trailer" keyword (Issue #58)
 - Fixed handling of invalid, but common, PDF files with a generation number of